back-end/requirements.txt
uvicorn==0.12.1
aiohttp==3.6.2
fastapi==0.61.1
databases==0.3.2
sqlalchemy==1.3.19
aiomysql==0.0.20
mysqlclient==2.0.1
pandas==1.1.2
xlrd==1.2.0
scikit-learn==0.23.2
aiofiles==0.5.0
jieba==0.42.1
pyyaml==5.3.1
networkx==2.5
torch==1.1.0
